+++ src/OpenTelemetry.Exporter.InfluxDB/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated OpenTelemetry core component version(s) to `1.15.1`.
+  ([#4020](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4020))
+
 * Custom metric tags are now exported.
   ([#4014](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4014))